refactor: rename folder to validators
